Update userclass field size in core plugins

+++ e107_plugins/poll/poll_sql.php
@@ -12,7 +12,7 @@ CREATE TABLE polls (
   poll_comment tinyint(1) unsigned NOT NULL default '1',
   poll_allow_multiple tinyint(1) unsigned NOT NULL default '0',
   poll_result_type tinyint(2) unsigned NOT NULL default '0',
-  poll_vote_userclass tinyint(3) unsigned NOT NULL default '0',
+  poll_vote_userclass smallint(5) unsigned NOT NULL default '0',
   poll_storage_method tinyint(1) unsigned NOT NULL default '0',
   PRIMARY KEY  (poll_id)
 ) TYPE=MyISAM;
